Delete the `reload` feature
The reload feature was implemented to clear the memorystore and reload
data from the previous configuration. But, `StringsLoader` is a
singleton. The previous configuration has to be kept in memory as long
as the application was running. This was memory consuming.

To replace the `reload` feature, `loadFromRemote` and `LoadFromLocal`
are now clearing the memory store before fetching any new data.

Both README and sample are updated according to this behaviour update.